backend/requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
openai
faiss-cpu
pypdf
opencv-python
numpy
python-dotenv
sentence-transformers
pymupdf
pdfplumber
torch